web/fix: project dashboard refactoring
