+++ packages/web/app/_components/project/embedded-pages/embedded-page-popover.tsx
@@ -4,8 +4,8 @@ import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
 import { EmbeddedPage } from '@prisma/client';
-import { PopoverMenuForm } from '../popover-menu-form';
-import { FormTextInput } from '../base/form/text-input';
+import { PopoverMenuForm } from '../../popover-menu-form';
+import { FormTextInput } from '../../base/form/text-input';
 import { ActionIcon } from '@mantine/core';
 import { IconTrash } from '@tabler/icons-react';
 import inputStyles from '@/styles/components/input.module.scss';
